requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==1.0.0
dj-database-url==1.2.0
